#3.9 naprawa crud (metody rud nie wywolują metody print)

+++ src/Main.java
@@ -25,6 +25,7 @@ public class Main {
 
         List<HostelFloor> hostelFloors = hostelFloorCRUD.readAllHostelFloors();
         System.out.println(hostelFloors);
+
         System.out.println("_________________________________________________");
         HostelFloor readtest1 = hostelFloorCRUD.readHostelFloor(1);
         HostelFloor readtest2 = hostelFloorCRUD.readHostelFloor(2);
@@ -34,16 +35,22 @@ public class Main {
 
 
         int updateAtId = 3;
-        HostelFloor hostelFloorToUpdate = hostelFloorCRUD.readHostelFloor(updateAtId);
-        hostelFloorToUpdate.setNumberOfFloors(8);
-        int index = hostelFloors.indexOf(hostelFloorToUpdate);
-        hostelFloorCRUD.updateHostelFloor(index, hostelFloorToUpdate);
+        HostelFloor hostelFloorToUpdate = null;
+        int index = 0;
+        if(hostelFloorCRUD.hostelFloorExists(updateAtId)){
+            hostelFloorToUpdate = hostelFloorCRUD.readHostelFloor(updateAtId);
+            hostelFloorToUpdate.setNumberOfFloors(8);
+            index = hostelFloors.indexOf(hostelFloorToUpdate);
+            hostelFloorCRUD.updateHostelFloor(index, hostelFloorToUpdate);
+        }else {hostelFloorCRUD.printHostelFloorNotFound(updateAtId);}
 
         int deleteAtId = 1;
-        //HostelFloor hostelFloorToDelete = hostelFloorCRUD.readHostelFloor(deleteAtId);
-        hostelFloorCRUD.deleteHostelFloor(deleteAtId);
-        System.out.println(hostelFloors);
-        //System.out.println(hostelFloorCRUD.getHostelFloorsID());
+        if(hostelFloorCRUD.hostelFloorExists(deleteAtId)){
+            //HostelFloor hostelFloorToDelete = hostelFloorCRUD.readHostelFloor(deleteAtId);
+            hostelFloorCRUD.deleteHostelFloor(deleteAtId);
+            System.out.println(hostelFloors);
+            //System.out.println(hostelFloorCRUD.getHostelFloorsID());
+        }else {hostelFloorCRUD.printHostelFloorNotFound(deleteAtId);}
 
         HostelFloor test5 = new HostelFloor(20, 2011, 200000, "usun", "Olejowa", "3b", 4, 30, false, "fajny budynek", 1, 15, false, false, 500, "Fajny pokoj");
         hostelFloorCRUD.createHostelFloor(test5);
@@ -51,18 +58,24 @@ public class Main {
         //System.out.println(hostelFloorCRUD.getHostelFloorsID());
 
         deleteAtId = 2;
-        hostelFloorCRUD.deleteHostelFloor(deleteAtId);
-        //System.out.println(hostelFloors);
+            if(hostelFloorCRUD.hostelFloorExists(deleteAtId)){
+            hostelFloorCRUD.deleteHostelFloor(deleteAtId);
+            //System.out.println(hostelFloors);
+        }else {hostelFloorCRUD.printHostelFloorNotFound(deleteAtId);}
 
         updateAtId = 1;
-        hostelFloorToUpdate = hostelFloorCRUD.readHostelFloor(updateAtId);
-        if(hostelFloorToUpdate != null){
+        if(hostelFloorCRUD.hostelFloorExists(updateAtId)){
+            hostelFloorToUpdate = hostelFloorCRUD.readHostelFloor(updateAtId);
             hostelFloorToUpdate.setNumberOfFloors(6);
             index = hostelFloors.indexOf(hostelFloorToUpdate);
             hostelFloorCRUD.updateHostelFloor(index, hostelFloorToUpdate);
-        }
+        }else {hostelFloorCRUD.printHostelFloorNotFound(updateAtId);}
 
-        HostelFloor readtest3 = hostelFloorCRUD.readHostelFloor(1);
-        if(readtest3 != null){System.out.println(readtest3);}
+
+        int readAtID = 1;
+            if(hostelFloorCRUD.hostelFloorExists(readAtID)){
+            HostelFloor readtest3 = hostelFloorCRUD.readHostelFloor(readAtID);
+            System.out.println(readtest3);
+        }else {hostelFloorCRUD.printHostelFloorNotFound(readAtID);}
     }
 }
